Implemented non-HTTP based import/export
