+++ Raven.Abstractions/Smuggler/ISmugglerApi.cs
@@ -1,0 +1,8 @@
+namespace Raven.Abstractions.Smuggler
+{
+	public interface ISmugglerApi
+	{
+		void ExportData(SmugglerOptions options, bool incremental);
+		void ImportData(SmugglerOptions options, bool incremental);
+	}
+}
