+++ Raven.Abstractions/Smuggler/SmugglerAction.cs
@@ -1,0 +1,8 @@
+namespace Raven.Abstractions.Smuggler
+{
+	public enum SmugglerAction
+	{
+		Import = 1,
+		Export,
+	}
+}
