--- Raven.Smuggler/ISmugglerApi.cs
@@ -1,8 +1,0 @@
-namespace Raven.Smuggler
-{
-	public interface ISmugglerApi
-	{
-		void ExportData(SmugglerOptions options, bool incremental);
-		void ImportData(SmugglerOptions options, bool incremental);
-	}
-}
