+++ Raven.Smuggler/Program.cs
@@ -8,6 +8,7 @@ using System.IO;
 using System.Net;
 using NDesk.Options;
 using Raven.Abstractions.Data;
+using Raven.Abstractions.Smuggler;
 
 namespace Raven.Smuggler
 {
