--- Raven.Smuggler/SmugglerAction.cs
@@ -1,8 +1,0 @@
-namespace Raven.Smuggler
-{
-	public enum SmugglerAction
-	{
-		Import = 1,
-		Export,
-	}
-}
