+++ Raven.Smuggler/SmugglerApi.cs
@@ -6,9 +6,6 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.Globalization;
-using System.IO;
-using System.IO.Compression;
 using System.Linq;
 using System.Net;
 using System.Text;
@@ -16,13 +13,29 @@ using Newtonsoft.Json;
 using Raven.Abstractions.Connection;
 using Raven.Abstractions.Data;
 using Raven.Abstractions.Extensions;
+using Raven.Abstractions.Smuggler;
 using Raven.Client.Extensions;
 using Raven.Json.Linq;
 
 namespace Raven.Smuggler
 {
-	public class SmugglerApi : ISmugglerApi
+	public class SmugglerApi : SmugglerApiBase
 	{
+		protected override RavenJArray GetIndexes(int totalCount)
+		{
+			RavenJArray indexes = null;
+			var request = CreateRequest("/indexes?pageSize=128&start=" + totalCount);
+			request.ExecuteRequest(reader => indexes = RavenJArray.Load(new JsonTextReader(reader)));
+			return indexes;
+		}
+
+		private static string StripQuotesIfNeeded(RavenJToken value)
+		{
+			var str = value.ToString(Formatting.None);
+			if (str.StartsWith("\"") && str.EndsWith("\""))
+				return str.Substring(1, str.Length - 2);
+			return str;
+		}
 		public RavenConnectionStringOptions ConnectionStringOptions { get; private set; }
 		private readonly HttpRavenRequestFactory httpRavenRequestFactory = new HttpRavenRequestFactory();
 
@@ -46,143 +59,15 @@ namespace Raven.Smuggler
 			return httpRavenRequestFactory.Create(builder.ToString(), method, ConnectionStringOptions);
 		}
 
-		public void ExportData(SmugglerOptions options, bool incremental = false)
+		protected override RavenJArray GetDocuments(Guid lastEtag)
 		{
-			var lastDocsEtag = Guid.Empty;
-			var lastAttachmentEtag = Guid.Empty;
-			var folder = options.File;
-			var etagFileLocation = Path.Combine(folder, "IncrementalExport.state.json");
-			if (incremental == true)
-			{
-				if (Directory.Exists(folder) == false)
-				{
-					Directory.CreateDirectory(folder);
-				}
-
-				options.File = Path.Combine(folder, DateTime.Now.ToString("yyyy-MM-dd-HH-mm", CultureInfo.InvariantCulture) + ".ravendb-incremental-dump");
-				if (File.Exists(options.File))
-				{
-					var counter = 1;
-					var found = false;
-					while (found == false)
-					{
-						options.File = Path.Combine(folder, DateTime.Now.ToString("yyyy-MM-dd-HH-mm", CultureInfo.InvariantCulture) + " - " + counter + ".ravendb-incremental-dump");
-
-						if (File.Exists(options.File) == false)
-							found = true;
-						counter++;
-					}
-				}
-
-				if (File.Exists(etagFileLocation))
-				{
-					using (var streamReader = new StreamReader(new FileStream(etagFileLocation, FileMode.Open)))
-					using (var jsonReader = new JsonTextReader(streamReader))
-					{
-						var ravenJObject = RavenJObject.Load(jsonReader);
-						lastDocsEtag = new Guid(ravenJObject.Value<string>("LastDocEtag"));
-						lastAttachmentEtag = new Guid(ravenJObject.Value<string>("LastAttachmentEtag"));
-					}
-				}
-			}
-
-
-			using (var streamWriter = new StreamWriter(new GZipStream(File.Create(options.File), CompressionMode.Compress)))
-			{
-				var jsonWriter = new JsonTextWriter(streamWriter)
-				{
-					Formatting = Formatting.Indented
-				};
-				jsonWriter.WriteStartObject();
-				jsonWriter.WritePropertyName("Indexes");
-				jsonWriter.WriteStartArray();
-				if (options.OperateOnTypes.HasFlag(ItemType.Indexes))
-				{
-					ExportIndexes(jsonWriter);
-				}
-				jsonWriter.WriteEndArray();
-
-				jsonWriter.WritePropertyName("Docs");
-				jsonWriter.WriteStartArray();
-				if (options.OperateOnTypes.HasFlag(ItemType.Documents))
-				{
-					lastDocsEtag = ExportDocuments(options, jsonWriter, lastDocsEtag);
-				}
-				jsonWriter.WriteEndArray();
-
-				jsonWriter.WritePropertyName("Attachments");
-				jsonWriter.WriteStartArray();
-				if (options.OperateOnTypes.HasFlag(ItemType.Attachments))
-				{
-					lastAttachmentEtag = ExportAttachments(jsonWriter, lastAttachmentEtag);
-				}
-				jsonWriter.WriteEndArray();
-
-				jsonWriter.WriteEndObject();
-				streamWriter.Flush();
-			}
-			if (incremental != true) 
-				return;
-
-			using (var streamWriter = new StreamWriter(File.Create(etagFileLocation)))
-			{
-				new RavenJObject
-				{
-					{"LastDocEtag", lastDocsEtag.ToString()},
-					{"LastAttachmentEtag", lastAttachmentEtag.ToString()}
-				}.WriteTo(new JsonTextWriter(streamWriter));
-				streamWriter.Flush();
-			}
+			RavenJArray documents = null;
+			var request = CreateRequest("/docs?pageSize=128&etag=" + lastEtag);
+			request.ExecuteRequest(reader => documents = RavenJArray.Load(new JsonTextReader(reader)));
+			return documents;
 		}
 
-		private void ExportIndexes(JsonTextWriter jsonWriter)
-		{
-			int totalCount = 0;
-			while (true)
-			{
-				RavenJArray indexes = null;
-				var request = CreateRequest("/indexes?pageSize=128&start=" + totalCount);
-				request.ExecuteRequest(reader => indexes = RavenJArray.Load(new JsonTextReader(reader)));
-
-				if (indexes.Length == 0)
-				{
-					Console.WriteLine("Done with reading indexes, total: {0}", totalCount);
-					break;
-				}
-				totalCount += indexes.Length;
-				Console.WriteLine("Reading batch of {0,3} indexes, read so far: {1,10:#,#;;0}", indexes.Length, totalCount);
-				foreach (RavenJToken item in indexes)
-				{
-					item.WriteTo(jsonWriter);
-				}
-			}
-		}
-
-		private Guid ExportDocuments(SmugglerOptions options, JsonTextWriter jsonWriter, Guid lastEtag)
-		{
-			int totalCount = 0;
-			while (true)
-			{
-				RavenJArray documents = null;
-				var request = CreateRequest("/docs?pageSize=128&etag=" + lastEtag);
-				request.ExecuteRequest(reader => documents = RavenJArray.Load(new JsonTextReader(reader)));
-
-				if (documents.Length == 0)
-				{
-					Console.WriteLine("Done with reading documents, total: {0}", totalCount);
-					return lastEtag;
-				}
-
-				var final = documents.Where(options.MatchFilters).ToList();
-				final.ForEach(item => item.WriteTo(jsonWriter));
-				totalCount += final.Count;
-
-				Console.WriteLine("Reading batch of {0,3} documents, read so far: {1,10:#,#;;0}", documents.Length, totalCount);
-				lastEtag = new Guid(documents.Last().Value<RavenJObject>("@metadata").Value<string>("@etag"));
-			}
-		}
-
-		private Guid ExportAttachments(JsonTextWriter jsonWriter, Guid lastEtag)
+		protected override Guid ExportAttachments(JsonTextWriter jsonWriter, Guid lastEtag)
 		{
 			int totalCount = 0;
 			while (true)
@@ -193,15 +78,15 @@ namespace Raven.Smuggler
 
 				if (attachmentInfo.Length == 0)
 				{
-					Console.WriteLine("Done with reading attachments, total: {0}", totalCount);
+					ShowProgress("Done with reading attachments, total: {0}", totalCount);
 					return lastEtag;
 				}
 
 				totalCount += attachmentInfo.Length;
-				Console.WriteLine("Reading batch of {0,3} attachments, read so far: {1,10:#,#;;0}", attachmentInfo.Length, totalCount);
+				ShowProgress("Reading batch of {0,3} attachments, read so far: {1,10:#,#;;0}", attachmentInfo.Length, totalCount);
 				foreach (var item in attachmentInfo)
 				{
-					Console.WriteLine("Downloading attachment: {0}", item.Value<string>("Key"));
+					ShowProgress("Downloading attachment: {0}", item.Value<string>("Key"));
 
 					byte[] attachmentData = null;
 					var requestData = CreateRequest("/static/" + item.Value<string>("Key"));
@@ -220,188 +105,43 @@ namespace Raven.Smuggler
 			}
 		}
 
-		public void ImportData(SmugglerOptions options, bool incremental = false)
+		protected override void PutAttachment(AttachmentExportInfo attachmentExportInfo)
 		{
-			if (incremental == false)
+			var request = CreateRequest("/static/" + attachmentExportInfo.Key, "PUT");
+			if (attachmentExportInfo.Metadata != null)
 			{
-				using (FileStream fileStream = File.OpenRead(options.File))
+				foreach (var header in attachmentExportInfo.Metadata)
 				{
-					ImportData(fileStream, options);
-				}
-				return;
-			}
-
-			var files = Directory.GetFiles(Path.GetFullPath(options.File))
-				.Where(file => ".ravendb-incremental-dump".Equals(Path.GetExtension(file), StringComparison.CurrentCultureIgnoreCase))
-				.OrderBy(File.GetLastWriteTimeUtc)
-				.ToArray();
-
-			if (files.Length == 0)
-				return;
-
-			var optionsWithoutIndexes = new SmugglerOptions
-			{
-				File = options.File,
-				Filters = options.Filters,
-				OperateOnTypes = options.OperateOnTypes & ~ItemType.Indexes
-			};
-
-			for (var i = 0; i < files.Length - 1; i++)
-			{
-				using (var fileStream = File.OpenRead(Path.Combine(options.File, files[i])))
-				{
-					ImportData(fileStream, optionsWithoutIndexes);
-				}
-			}
-
-			using (var fileStream = File.OpenRead(Path.Combine(options.File, files.Last())))
-			{
-				ImportData(fileStream, options);
-			}
-		}
-
-		public void ImportData(Stream stream, SmugglerOptions options, bool importIndexes = true)
-		{
-			EnsureDatabaseExists();
-
-			var sw = Stopwatch.StartNew();
-			// Try to read the stream compressed, otherwise continue uncompressed.
-			JsonTextReader jsonReader;
-			try
-			{
-				var streamReader = new StreamReader(new GZipStream(stream, CompressionMode.Decompress));
-
-				jsonReader = new JsonTextReader(streamReader);
-
-				if (jsonReader.Read() == false)
-					return;
-			}
-			catch (InvalidDataException)
-			{
-				stream.Seek(0, SeekOrigin.Begin);
-
-				var streamReader = new StreamReader(stream);
-
-				jsonReader = new JsonTextReader(streamReader);
-
-				if (jsonReader.Read() == false)
-					return;
-			}
-
-			if (jsonReader.TokenType != JsonToken.StartObject)
-				throw new InvalidDataException("StartObject was expected");
-
-			// should read indexes now
-			if (jsonReader.Read() == false)
-				return;
-			if (jsonReader.TokenType != JsonToken.PropertyName)
-				throw new InvalidDataException("PropertyName was expected");
-			if (Equals("Indexes", jsonReader.Value) == false)
-				throw new InvalidDataException("Indexes property was expected");
-			if (jsonReader.Read() == false)
-				return;
-			if (jsonReader.TokenType != JsonToken.StartArray)
-				throw new InvalidDataException("StartArray was expected");
-
-			while (jsonReader.Read() && jsonReader.TokenType != JsonToken.EndArray)
-			{
-				var index = RavenJToken.ReadFrom(jsonReader);
-				if (options.OperateOnTypes.HasFlag(ItemType.Indexes) == false)
-					continue;
-				var indexName = index.Value<string>("name");
-				if (indexName.StartsWith("Raven/") || indexName.StartsWith("Temp/"))
-					continue;
-
-				var request = CreateRequest("/indexes/" + indexName, "PUT");
-				request.Write(index.Value<RavenJObject>("definition"));
-				request.ExecuteRequest();
-
-			}
-
-			// should read documents now
-			if (jsonReader.Read() == false)
-				return;
-			if (jsonReader.TokenType != JsonToken.PropertyName)
-				throw new InvalidDataException("PropertyName was expected");
-			if (Equals("Docs", jsonReader.Value) == false)
-				throw new InvalidDataException("Docs property was expected");
-			if (jsonReader.Read() == false)
-				return;
-			if (jsonReader.TokenType != JsonToken.StartArray)
-				throw new InvalidDataException("StartArray was expected");
-			var batch = new List<RavenJObject>();
-			int totalCount = 0;
-			while (jsonReader.Read() && jsonReader.TokenType != JsonToken.EndArray)
-			{
-				var document = (RavenJObject)RavenJToken.ReadFrom(jsonReader);
-				if (options.OperateOnTypes.HasFlag(ItemType.Documents) == false)
-					continue;
-				if (options.MatchFilters(document) == false)
-					continue;
-
-				totalCount += 1;
-				batch.Add(document);
-				if (batch.Count >= 128)
-					FlushBatch(batch);
-			}
-			FlushBatch(batch);
-
-			var attachmentCount = 0;
-			if (jsonReader.Read() == false || jsonReader.TokenType == JsonToken.EndObject)
-				return;
-			if (jsonReader.TokenType != JsonToken.PropertyName)
-				throw new InvalidDataException("PropertyName was expected");
-			if (Equals("Attachments", jsonReader.Value) == false)
-				throw new InvalidDataException("Attachment property was expected");
-			if (jsonReader.Read() == false)
-				return;
-			if (jsonReader.TokenType != JsonToken.StartArray)
-				throw new InvalidDataException("StartArray was expected");
-			while (jsonReader.Read() && jsonReader.TokenType != JsonToken.EndArray)
-			{
-				attachmentCount += 1;
-				var item = RavenJToken.ReadFrom(jsonReader);
-				if (options.OperateOnTypes.HasFlag(ItemType.Attachments) == false)
-					continue;
-				var attachmentExportInfo =
-					new JsonSerializer
+					switch (header.Key)
 					{
-						Converters = { new TrivialJsonToJsonJsonConverter() }
-					}.Deserialize<AttachmentExportInfo>(new RavenJTokenReader(item));
-				Console.WriteLine("Importing attachment {0}", attachmentExportInfo.Key);
-
-				var request = CreateRequest("/static/" + attachmentExportInfo.Key, "PUT");
-				if (attachmentExportInfo.Metadata != null)
-				{
-					foreach (var header in attachmentExportInfo.Metadata)
-					{
-						switch (header.Key)
-						{
-							case "Content-Type":
-								request.WebRequest.ContentType = header.Value.Value<string>();
-								break;
-							default:
-								request.WebRequest.Headers.Add(header.Key, StripQuotesIfNeeded(header.Value));
-								break;
-						}
+						case "Content-Type":
+							request.WebRequest.ContentType = header.Value.Value<string>();
+							break;
+						default:
+							request.WebRequest.Headers.Add(header.Key, StripQuotesIfNeeded(header.Value));
+							break;
 					}
 				}
-
-				request.Write(attachmentExportInfo.Data);
-				request.ExecuteRequest();
 			}
-			Console.WriteLine("Imported {0:#,#;;0} documents and {1:#,#;;0} attachments in {2:#,#;;0} ms", totalCount, attachmentCount, sw.ElapsedMilliseconds);
+
+			request.Write(attachmentExportInfo.Data);
+			request.ExecuteRequest();
+
 		}
 
-		private static string StripQuotesIfNeeded(RavenJToken value)
+		protected override void PutIndex(string indexName, RavenJToken index)
 		{
-			var str = value.ToString(Formatting.None);
-			if (str.StartsWith("\"") && str.EndsWith("\""))
-				return str.Substring(1, str.Length - 2);
-			return str;
+			var request = CreateRequest("/indexes/" + indexName, "PUT");
+			request.Write(index.Value<RavenJObject>("definition"));
+			request.ExecuteRequest();
 		}
 
-		private void FlushBatch(List<RavenJObject> batch)
+		protected override void ShowProgress(string format, params object[] args)
+		{
+			Console.WriteLine(format, args);
+		}
+
+		protected override void FlushBatch(List<RavenJObject> batch)
 		{
 			var sw = Stopwatch.StartNew();
 
@@ -411,33 +151,25 @@ namespace Raven.Smuggler
 				var metadata = doc.Value<RavenJObject>("@metadata");
 				doc.Remove("@metadata");
 				commands.Add(new RavenJObject
-							    {
-							        {"Method", "PUT"},
-							        {"Document", doc},
-							        {"Metadata", metadata},
-							        {"Key", metadata.Value<string>("@id")}
-							    });
+								{
+									{"Method", "PUT"},
+									{"Document", doc},
+									{"Metadata", metadata},
+									{"Key", metadata.Value<string>("@id")}
+								});
 			}
 
 			var request = CreateRequest("/bulk_docs", "POST");
 			request.Write(commands);
 			request.ExecuteRequest();
 
-			Console.Write("Wrote {0} documents in {1}", batch.Count, sw.ElapsedMilliseconds);
+			ShowProgress("Wrote {0} documents in {1}", batch.Count, sw.ElapsedMilliseconds);
 
-			Console.WriteLine(" in {0:#,#;;0} ms", sw.ElapsedMilliseconds);
+			ShowProgress(" in {0:#,#;;0} ms", sw.ElapsedMilliseconds);
 			batch.Clear();
 		}
 
-		private class AttachmentExportInfo
-		{
-			public byte[] Data { get; set; }
-			public RavenJObject Metadata { get; set; }
-			public string Key { get; set; }
-		}
-
-		private bool ensuredDatabaseExists;
-		private void EnsureDatabaseExists()
+		protected override void EnsureDatabaseExists()
 		{
 			if (ensuredDatabaseExists ||
 				string.IsNullOrWhiteSpace(ConnectionStringOptions.DefaultDatabase))
