+++ Raven.Tests/MailingList/MapReduceIssue/CanPageThroughReduceResults.cs
@@ -1,4 +1,5 @@
 using Raven.Abstractions.Data;
+using Raven.Abstractions.Smuggler;
 using Raven.Client.Document;
 using Raven.Client.Linq;
 using Raven.Smuggler;
